+++ src/misc/xmlwrapp/node.cpp
@@ -267,7 +267,8 @@ xml::node::node (const char *name, const char *content) {
 xml::node::node (cdata cdata_info) {
     std::auto_ptr<node_impl> ap(pimpl_ = new node_impl);
 
-    if ( (pimpl_->xmlnode_ = xmlNewCDataBlock(0, reinterpret_cast<const xmlChar*>(cdata_info.t), std::strlen(cdata_info.t))) == 0) {
+    if ( (pimpl_->xmlnode_ = xmlNewCDataBlock(0, reinterpret_cast<const xmlChar*>(cdata_info.t), 
+                                              static_cast<int>(std::strlen(cdata_info.t)))) == 0) {
 	throw std::bad_alloc();
     }
